move ApiKeyEncryption to auth-core
